Move ConfigChanged out of RenderBase
There is this nice VideoConfig file that's perfect for it
